+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-No unreleased changes yet. 
+### Added
+- [#85](https://github.com/startable/pdtable/issues/85) CSV and Excel writers can write transposed tables.
 
 ## [0.0.7] - 2021-01-18